--- source/server/src/main/java/registry/exceptions/ItemNotFoundException.java
@@ -1,15 +1,0 @@
-package registry.exceptions;
-
-public class ItemNotFoundException extends NotFoundException {
-    public ItemNotFoundException() { super(); }
-
-    public ItemNotFoundException(String message) { super(message); }
-
-    public ItemNotFoundException(String message, Throwable cause) { super(message, cause); }
-
-    public ItemNotFoundException(Throwable cause) { super(cause); }
-
-    public ItemNotFoundException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-        super(message, cause, enableSuppression, writableStackTrace);
-    }
-}
