--- source/server/src/main/java/registry/exceptions/ReportNotFoundException.java
@@ -1,5 +1,0 @@
-package registry.exceptions;
-
-public class ReportNotFoundException extends NotFoundException{
-
-}
