+++ source/tests/integrationTests/build.gradle
@@ -35,6 +35,7 @@ dependencies {
     testCompile project(':server')
     testCompile 'org.mockito:mockito-all:1.9.5'
     testCompile 'org.junit.jupiter:junit-jupiter:5.4.0'
+    testCompile group: 'org.junit.platform', name: 'junit-platform-runner', version: '1.5.2'
     testImplementation 'org.springframework.boot:spring-boot-starter-test:2.0.5.RELEASE'
 }
